+++ src/components.d.ts
@@ -12,7 +12,17 @@ export namespace Components {
   interface PhemiumCard {
     'card': any;
     'config': { apiEndpoint: string; token: string; userId: number; hideSubmitButton: boolean; selectionStyle: string; showStaticText: boolean; soloText: boolean; submitButtonText: string; maxFileSize: any; formStyle: string; inputChecked: boolean; };
+    /**
+    * Function to handle submit event when user finishes inserting values. It uploads resources if needed and emits an event with   an array containing all form values on it, prepared to send to phemium.
+    */
+    'handleSubmit': () => Promise<void>;
     'language': string;
+  }
+  interface PhemiumPush {
+    'active': boolean;
+    'customHandler': boolean;
+    'initialize': (phemiumConfig: any, firebaseConfig: any, appID: string) => Promise<void>;
+    'showNotification': boolean;
   }
 }
 
@@ -24,8 +34,15 @@ declare global {
     prototype: HTMLPhemiumCardElement;
     new (): HTMLPhemiumCardElement;
   };
+
+  interface HTMLPhemiumPushElement extends Components.PhemiumPush, HTMLStencilElement {}
+  var HTMLPhemiumPushElement: {
+    prototype: HTMLPhemiumPushElement;
+    new (): HTMLPhemiumPushElement;
+  };
   interface HTMLElementTagNameMap {
     'phemium-card': HTMLPhemiumCardElement;
+    'phemium-push': HTMLPhemiumPushElement;
   }
 }
 
@@ -41,9 +58,16 @@ declare namespace LocalJSX {
     'onShowInformation'?: (event: CustomEvent<any>) => void;
     'onUploadingFiles'?: (event: CustomEvent<any>) => void;
   }
+  interface PhemiumPush extends JSXBase.HTMLAttributes<HTMLPhemiumPushElement> {
+    'active'?: boolean;
+    'customHandler'?: boolean;
+    'onOnNotification'?: (event: CustomEvent<any>) => void;
+    'showNotification'?: boolean;
+  }
 
   interface IntrinsicElements {
     'phemium-card': PhemiumCard;
+    'phemium-push': PhemiumPush;
   }
 }